requirements.txt
flask
redis
requests
flasgger
python-dotenv
openai
google-generativeai